// types/video.ts
export type Video = {
  id: string;
  slug: string;
  title: string;
  description: string;
  youtube_url: string;
  youtube_id: string;
  thumbnail_url: string;
  subject: string;
  grade_level: string;
  topic: string;
  instructor: string;
  duration: string;
  language: string;
  published: boolean;
  featured: boolean;
  display_order: string;
  published_date: string;
};